Major Update to Linear Actuators
Renamed the LinearActuator class to SabretoothLinearActuator.

Added LinearActuator function for legacy support. The LinearActuator
function returns a SabretoothLinearActuator object.

Added the invert() method to the SabretoothLinearActuator class,
allowing the extend and retract values to be swapped without having to
make a new object.

Added new class ActuatorGroup. ActuatorGroup takes in a list, tuple, or
dictionary of actuator objects and allows them to be manipulated as a
group or individually